fix: split FSM and KnxClient
